+++ ui/pages/defi/components/defi-details-page.tsx
@@ -23,7 +23,7 @@ import { DEFAULT_ROUTE } from '../../../helpers/constants/routes';
 
 import { getPreferences, getSelectedAccount } from '../../../selectors';
 import { CHAIN_IDS } from '../../../../shared/constants/network';
-import { useFormatters } from '../../../helpers/formatters';
+import { useFormatters } from '../../../hooks/formatters';
 import { AssetCellBadge } from '../../../components/app/assets/asset-list/cells/asset-cell-badge';
 import { getDefiPositions } from '../../../selectors/assets';
 import DefiDetailsList, {
